Project Dice Game
Here is a play on the dice game Tenzies. I have built this project using React and have included several features such as betting on outcomes, use of derived state to have my die components be able to be held by the user. I have also implemented some useEffects to handle some of the side effects required for this app such as monitoring for a met Win condition.